Fixed type of player pref
changed from getfloat to getint

+++ Assets/Scripts/UI Scripts/Main Menu/MainMenuController.cs
@@ -109,7 +109,7 @@ public class MainMenuController : MonoSingleton<MainMenuController>
             case eMainMenuCanvas.Main:
                 //MAIN MENU
                 MainMenuCanvas.gameObject.SetActive(true);
-                if (PlayerPrefs.GetFloat("TutorialAvailable") == 0)//NEW CODE
+                if (PlayerPrefs.GetInt("TutorialAvailable") == 0)//NEW CODE
                 {
                     TutorialsCanvas.gameObject.SetActive(true);//NEW CODE
                 }
